// webapp/src/@generated/schemas/visitorDto.ts
/**
 * Generated by orval v6.12.1 🍺
 * Do not edit manually.
 * Vol
 * Registration api
 * OpenAPI spec version: 0.1
 */
import type { VisitorDtoStatus } from './visitorDtoStatus';
import type { VisitorDtoGender } from './visitorDtoGender';
import type { VisitorDtoVisitsItem } from './visitorDtoVisitsItem';

export interface VisitorDto {
  id: string;
  name: string;
  birthday: string;
  hasDisability: boolean;
  status: VisitorDtoStatus;
  gender: VisitorDtoGender;
  visits?: VisitorDtoVisitsItem[];
}